refactor(examples): 迁移 selection 至 plugins 目录
