+++ examples/src/pages/plugins/selection/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Graph, Keyboard, Selection } from '@antv/x6'
-import '../index.less'
+import '../../index.less'
 
 export class SelectionExample extends React.Component {
   private container!: HTMLDivElement
